Create, use ColoredMesh ctor
Moves some logic into ColoredMesh from Forms to reduce code repetition.
